Use float

+++ nettigo_air_monitor/model.py
@@ -29,7 +29,7 @@ class ConnectionOptions:
 class NAMSensors:
     """Data class for NAM sensors."""
 
-    bh1750_illuminance: int | None
+    bh1750_illuminance: float | None
     bme280_humidity: float | None
     bme280_pressure: float | None
     bme280_temperature: float | None
